flask_resource_api/requirements.txt
gunicorn
Flask
pymongo
bcrypt
pyjwt[crypto]
schedule
requests